+++ model/src/main/java/com/aplana/sbrf/taxaccounting/model/QNotification.java
@@ -5,14 +5,13 @@ import static com.querydsl.core.types.PathMetadataFactory.*;
 import com.querydsl.core.types.dsl.*;
 
 import com.querydsl.core.types.PathMetadata;
-
 import javax.annotation.Generated;
-
 import com.querydsl.core.types.Path;
 
 import com.querydsl.sql.ColumnMetadata;
-
 import java.sql.Types;
+
+
 
 
 /**
@@ -47,7 +46,7 @@ public class QNotification extends com.querydsl.sql.RelationalPathBase<QNotifica
 
     public final StringPath text = createString("text");
 
-    public final NumberPath<Byte> type = createNumber("type", Byte.class);
+    public final NumberPath<Integer> type = createNumber("type", Integer.class);
 
     public final NumberPath<Integer> userId = createNumber("userId", Integer.class);
